+++ public/literals/english/common.json
@@ -5,6 +5,7 @@
   "---author": "",
   "gabriel": "Romay Machado, Gabriel",
   "cv": "https://www.linkedin.com/in/gabriel-romay-machado-40050a114/?locale=en_US",
-  "---recipeListTemplate": "",
-  "participants": "Participants"
+  "---about": "",
+  "participants": "Participants",
+  "about": "About"
 }
