+++ public/literals/japanese/common.json
@@ -6,5 +6,6 @@
   "gabriel": "ロメイ　マーチャード　ガブリエル",
   "cv": "https://www.linkedin.com/in/gabriel-romay-machado-40050a114/?locale=en_US",
   "---recipeListTemplate": "",
-  "participants": "Participants"
+  "participants": "Participants",
+  "about": "About"
 }
