+++ public/literals/spanish/common.json
@@ -5,6 +5,7 @@
   "---author": "",
   "gabriel": "Gabriel Romay Machado",
   "cv": "https://www.linkedin.com/in/gabriel-romay-machado-40050a114/",
-  "---recipeListTemplate": "",
-  "participants": "Participantes"
+  "---about": "",
+  "participants": "Participantes",
+  "about": "Acerca de"
 }
