+++ src/App.tsx
@@ -9,19 +9,23 @@ import {
 import LanguageProvider from './language/languageProvider';
 
 import NotFoundPage from './pages/notFound';
-import RecipeDetailPage from './pages/recipeDetail';
-import RecipeListPage from './pages/recipeList';
+import AboutPage from './pages/about';
+import DetailPage from './pages/recipeDetail';
+import ListPage from './pages/recipeList';
 
 export default hot(module)(() => (
   <LanguageProvider>
     <div className="App">
       <Router>
         <Switch>
+          <Route exact path="/about">
+            <AboutPage />
+          </Route>
           <Route exact path="/list">
-            <RecipeListPage />
+            <ListPage />
           </Route>
           <Route exact path="/detail/:recipeId">
-            <RecipeDetailPage />
+            <DetailPage />
           </Route>
           <Redirect from="/" exact to="/list" />
           <Route component={NotFoundPage} />
