+++ src/pages/recipeList/index.tsx
@@ -12,7 +12,6 @@ export default () => {
   const common = translate('common');
   const recipe = translate('recipe');
   const recipesNamesLiterals = translate('recipesNames');
-  const projectData = translate('projectData');
 
   const literals = { ...common, ...recipe };
   const getLiterals = ({ difficulty, preparationTime, howManyIngredients, participants }: Literals) => ({
@@ -49,17 +48,6 @@ export default () => {
   return recipes &&
     <RecipeListTemplate
       literals={getLiterals(literals)}
-      authorData={{
-        name: common.gabriel,
-        email: 'romay.gabriel@gmail.com',
-        url: common.cv,
-      }}
-      projectData={{
-        description: projectData.description,
-        url: 'https://github.com/W01fw00d/cooking-with-amateurs/blob/master/README.md',
-        participants:
-          ["Mateu", "Marina", "Jose Luis", "Iris", "Malén", "PJ", "Francina", "Lluis", "Marta", "Gabriel"],
-      }}
       search={search}
       itemList={recipes}
       languageData={{
